Add hire trader dialog on homepage
