+++ src/app/page.tsx
@@ -16,6 +16,7 @@ import {
 } from "@/components/feed-line";
 import { PendingApprovalCard } from "@/components/pending-approval-card";
 import { ConvexIdentityDebug } from "@/components/convex-identity-debug";
+import { TraderCreationDialog } from "@/components/trader-creation-flow";
 import type { AgentActivity } from "@/hooks/use-agent";
 import { useActivityFeed } from "@/hooks/use-activity-feed";
 import {
@@ -186,6 +187,7 @@ function Dashboard({ displayName }: { displayName: string }) {
     traderId: string;
     dealId: string | null;
   } | null>(null);
+  const [hireDialogOpen, setHireDialogOpen] = useState(false);
 
   const activity = useMemo(() => feedData?.activity ?? [], [feedData]);
   const traderNames = feedData?.traderNames ?? {};
@@ -239,6 +241,7 @@ function Dashboard({ displayName }: { displayName: string }) {
             portfolioLoading={portfolioLoading}
             traderFilter={traderFilter}
             onTraderFilter={setTraderFilter}
+            onHireTrader={() => setHireDialogOpen(true)}
           />
           <TraderFeedPanel
             activity={filteredActivity}
@@ -270,6 +273,10 @@ function Dashboard({ displayName }: { displayName: string }) {
         }}
         traderId={approvalCtx?.traderId ?? null}
         dealId={approvalCtx?.dealId ?? null}
+      />
+      <TraderCreationDialog
+        open={hireDialogOpen}
+        onOpenChange={setHireDialogOpen}
       />
     </div>
   );
@@ -549,12 +556,14 @@ function TradingDeskPanel({
   portfolioLoading,
   traderFilter,
   onTraderFilter,
+  onHireTrader,
 }: {
   nowMs: number;
   portfolio: Portfolio | undefined;
   portfolioLoading: boolean;
   traderFilter: string | null;
   onTraderFilter: (id: string | null) => void;
+  onHireTrader: () => void;
 }) {
   const traders = portfolio?.traders ?? [];
 
@@ -564,12 +573,13 @@ function TradingDeskPanel({
         title="Your Trading Desk"
         meta={`${traders.length} TRADER${traders.length === 1 ? "" : "S"}`}
         action={
-          <Link
-            href="/traders/new"
+          <button
+            type="button"
+            onClick={onHireTrader}
             className="border border-[var(--t-divider)] px-2 py-1 text-[10px] uppercase tracking-wider text-[var(--t-accent)] hover:border-[var(--t-accent)]"
           >
-            Hire
-          </Link>
+            Hire Trader
+          </button>
         }
       />
 
@@ -583,12 +593,13 @@ function TradingDeskPanel({
           <p className="text-sm uppercase tracking-wider text-[var(--t-muted)]">
             No traders on your desk
           </p>
-          <Link
-            href="/traders/new"
+          <button
+            type="button"
+            onClick={onHireTrader}
             className="mt-4 inline-block border border-[var(--t-accent)] px-5 py-2 text-xs uppercase tracking-wider text-[var(--t-accent)] hover:bg-[var(--t-accent-soft)]"
           >
-            Open hiring file
-          </Link>
+            Hire Trader
+          </button>
         </div>
       )}
       {!portfolioLoading && traders.length > 0 && (
